+++ ref/hash_shake.c
@@ -38,7 +38,7 @@ void gen_message_random(unsigned char *R, const unsigned char *sk_prf,
     shake256_inc_init(&s_inc);
     shake256_inc_absorb(&s_inc, sk_prf, SPX_N);
     shake256_inc_absorb(&s_inc, optrand, SPX_N);
-    shake256_inc_absorb(&s_inc, m, mlen);
+    shake256_inc_absorb(&s_inc, m, (size_t)mlen);
     shake256_inc_finalize(&s_inc);
     shake256_inc_squeeze(R, SPX_N, &s_inc);
     shake256_inc_ctx_release(&s_inc);
@@ -68,7 +68,7 @@ void hash_message(unsigned char *digest, uint64_t *tree, uint32_t *leaf_idx,
     shake256_inc_init(&s_inc);
     shake256_inc_absorb(&s_inc, R, SPX_N);
     shake256_inc_absorb(&s_inc, pk, SPX_PK_BYTES);
-    shake256_inc_absorb(&s_inc, m, mlen);
+    shake256_inc_absorb(&s_inc, m, (size_t)mlen);
     shake256_inc_finalize(&s_inc);
     shake256_inc_squeeze(buf, SPX_DGST_BYTES, &s_inc);
     shake256_inc_ctx_release(&s_inc);
